+++ lily-design-system-angular-text-size-select/index.ts
@@ -1,1 +1,8 @@
-export { TextSizeSelect } from "./text-size-select.component";
+export {
+  TextSizeSelect,
+  TextSizeSelectIcon,
+  LATIN_CAPITAL_LETTER_A,
+  nextTextSizeSelectId,
+  sizeName,
+} from "./text-size-select.component";
+export type { ChildArgs } from "./text-size-select.component";